[OPEN-569] [OPEN-571] Sandbox projects (#551)
This PR adds the creation of Sandbox Projects when signing up for the
Console.

This is done by updating the existing `OnboardingCreateProject` rpc in
the Intermediate Service. The corresponding
`createProjectForCurrentUser` store function no longer creates a `dev`
and `prod` Project, and instead creates a single Project using the
`displayName` and `appUrl` provided by the client.

---

In order to allow for advancement beyond the Sandbox Project, this PR
also adds support for creating new Projects from the Console.

This is done via a new `ConsoleCreateProject` rpc in the Backend
Service. This RPC works very similarly to its analog in the Intermediate
Service with the main difference being that it returns the Project and
backing Organization ID instead of creating a new session. Once invoked,
the client redirects the user to the `/switch-organization` routed using
the returned Organization ID, which works like usual, so the user is
given a new session for the new Project and redirected back to the
Console.

<img width="1659" height="948" alt="Screenshot 2025-07-17 at 3 56 30 PM"
src="https://github.com/user-attachments/assets/41bda257-1345-4771-9120-0db754e71eb9"
/>
<img width="1659" height="949" alt="Screenshot 2025-07-17 at 3 56 41 PM"
src="https://github.com/user-attachments/assets/9dbd9301-449a-40c8-b785-35477ab26879"
/>