[OPEN-587] Rename `dogfood` to `console` (#553)
This PR updates all references to `DOGFOOD` to `CONSOLE`, `Dogfood` to
`Console`, and `dogfood` to `console`. This is across all services and
static files.

This is in preparation for a the console configuration discovery PR to
minimize the amount of double work required as we roll out self-hosting
support.

Tested against `dev1`.

NOTE: This change requires environment variable updates via terraform.
These changes have already been applied in all environments, so the
service is safe to release.