[OPEN-594] Load console vault domain through static file (#557)
This PR moves how the Console loads its vault domain from being based on
esbuild to instead loading from a static file at `/config.json`.

This is in service of making self-hosting easier, by removing a
configuration-time dependency on esbuild. We can now ship a mostly
complete build of the Console, and the only final step is serving a
correct `/config.json`, which can easily happen without a bundler.