[OPEN-606] Move Login flow query param state into localStorage (#564)
This PR moves the login flow's URL-supplied state
(`relayed-session-state`, `redirect-uri`, and
`return-relayed-session-token-as-query-param`) to persist in
localstorage.

Today, this state is only loaded from the URL, but we also strip it from
the URL upon loading it. That interacts poorly with browser history --
certain patterns of hitting back or refreshing can lose that state.

With this PR, this state continues to be accepted as a query parameter,
but is persisted in local storage. Finishing a login resets the state.