[OPEN-586] Add update-project ctl command + bypass s3 with direct urls (#562)
This PR adds an `update-project` command to the ctl tool, which allows
self-hosters to directly control, through the database, a Project's
Vault domain and logos. This allows self-hosters to conveniently bypass
dependencies on S3 and Cloudflare.

For logos, a new set of columns for `logo_url` and `dark_mode_logo_url`
take precedence over other settings. If enabled, these are
unconditionally used as the hotlinkable URL to use for presentational
purposes.